Bug 1580687 - cleanup and address PR comments
