+++ treeherder/webapp/api/performance_data.py
@@ -5,6 +5,7 @@ from collections import defaultdict
 import django_filters
 from django.conf import settings
 from django.db import transaction
+from django.db.models import Count
 from rest_framework import (exceptions,
                             filters,
                             generics,
@@ -32,7 +33,10 @@ from .performance_serializers import (IssueTrackerSerializer,
                                       PerformanceBugTemplateSerializer,
                                       PerformanceFrameworkSerializer,
                                       PerformanceQueryParamsSerializer,
-                                      PerformanceSummarySerializer, ValidityDashboardParamsSerializer)
+                                      PerformanceSummarySerializer,
+                                      TestSuiteHealthParamsSerializer,
+                                      TestSuiteHealthSerializer)
+from .utils import GroupConcat
 
 
 class PerformanceSignatureViewSet(viewsets.ViewSet):
@@ -493,49 +497,24 @@ class PerformanceSummary(generics.ListAPIView):
         serializer = self.get_serializer(self.queryset, many=True)
         return Response(data=serializer.data)
 
-from django.db.models import Aggregate, CharField, Count
 
-
-class GroupConcat(Aggregate):
-    function = 'GROUP_CONCAT'
-    template = '%(function)s(%(distinct)s%(expressions)s)'
-    allow_distinct = True
-
-    def __init__(self, expression, distinct=False, **extra):
-        super().__init__(
-            expression,
-            distinct='DISTINCT ' if distinct else '',
-            output_field=CharField(),
-            **extra)
-
-
-class ValidityDashboardViewSet(viewsets.ViewSet):
+class TestSuiteHealthViewSet(viewsets.ViewSet):
 
     def list(self, request):
-        query_params = ValidityDashboardParamsSerializer(data=request.query_params)
+        query_params = TestSuiteHealthParamsSerializer(data=request.query_params)
         if not query_params.is_valid():
             return Response(data=query_params.errors,
                             status=HTTP_400_BAD_REQUEST)
 
         framework_id = query_params.validated_data['framework']
-
-        query_set_one = (PerformanceSignature.objects
+        query_set = (PerformanceSignature.objects
                      .prefetch_related('performancealert')
+                     .filter(framework_id=framework_id, parent_signature_id=None)
                      .values('suite', 'test')
                      .annotate(repositories=GroupConcat('repository_id', distinct=True))
                      .annotate(platforms=GroupConcat('platform_id', distinct=True))
                      .annotate(total_alerts=Count('performancealert'))
-                     .filter(framework_id=framework_id)
                      .order_by('suite', 'test'))
 
-        # change in place
-        aggregated_signature_data = list(query_set_one)
-        # data_two = list(query_set_two)
-        for item in aggregated_signature_data:
-            item['repositories'] = self.__list_form(item['repositories'])
-            item['platforms'] = self.__list_form(item['platforms'])
-
-        return Response(aggregated_signature_data)
-
-    def __list_form(self, comma_strings):
-        return comma_strings.split(',')
+        serializer = TestSuiteHealthSerializer(query_set, many=True)
+        return Response(data=serializer.data)
