+++ treeherder/webapp/api/urls.py
@@ -122,8 +122,8 @@ default_router.register(r'performance/issue-tracker',
                         performance_data.PerformanceIssueTrackerViewSet,
                         base_name='performance-issue-tracker')
 default_router.register(r'performance/validity-dashboard',
-                        performance_data.ValidityDashboardViewSet,
-                         base_name='validity-dashboard')
+                        performance_data.TestSuiteHealthViewSet,
+                        base_name='validity-dashboard')
 default_router.register(r'bugzilla', bugzilla.BugzillaViewSet,
                         base_name='bugzilla')
 default_router.register(r'jobdetail', jobs.JobDetailViewSet,
